backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
google-genai==1.51.0
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.12.4
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
httpx==0.28.1

